+++ Source/Alternet.UI.Pal/Api/Font.Api.h
@@ -17,7 +17,7 @@ ALTERNET_UI_API char16_t* Font_GetName_(Font* obj)
     return AllocPInvokeReturnString(obj->GetName());
 }
 
-ALTERNET_UI_API float Font_GetSizeInPoints_(Font* obj)
+ALTERNET_UI_API double Font_GetSizeInPoints_(Font* obj)
 {
     return obj->GetSizeInPoints();
 }
@@ -52,7 +52,7 @@ ALTERNET_UI_API void Font_CloseFamiliesArray_(void* array)
     Font::CloseFamiliesArray(array);
 }
 
-ALTERNET_UI_API void Font_Initialize_(Font* obj, GenericFontFamily genericFamily, const char16_t* familyName, float emSizeInPoints, FontStyle style)
+ALTERNET_UI_API void Font_Initialize_(Font* obj, GenericFontFamily genericFamily, const char16_t* familyName, double emSizeInPoints, FontStyle style)
 {
     obj->Initialize(genericFamily, ToOptional(familyName), emSizeInPoints, style);
 }
