+++ Source/Alternet.UI/Handlers/Layout/StackPanel/StackPanelHandler.HorizontalLayout.cs
@@ -15,8 +15,8 @@ namespace Alternet.UI
             {
                 var stackPanelPadding = Control.Padding;
 
-                float width = 0;
-                float maxHeight = 0;
+                double width = 0;
+                double maxHeight = 0;
                 foreach (var control in Handler.AllChildrenIncludedInLayout)
                 {
                     var margin = control.Margin;
@@ -25,14 +25,14 @@ namespace Alternet.UI
                     maxHeight = Math.Max(maxHeight, preferredSize.Height + margin.Vertical);
                 }
 
-                return new Size(width + stackPanelPadding.Horizontal, float.IsNaN(Control.Height) ? maxHeight + stackPanelPadding.Vertical : Control.Height);
+                return new Size(width + stackPanelPadding.Horizontal, double.IsNaN(Control.Height) ? maxHeight + stackPanelPadding.Vertical : Control.Height);
             }
 
             public override void Layout()
             {
                 var childrenLayoutBounds = Handler.ChildrenLayoutBounds;
 
-                float x = 0;
+                double x = 0;
                 foreach (var control in Handler.AllChildrenIncludedInLayout)
                 {
                     var margin = control.Margin;
