+++ Source/NativeApi/NativeApi/Api/Pen/Pen.cs
@@ -5,6 +5,6 @@ namespace NativeApi.Api
 {
     public class Pen
     {
-        public void Initialize(PenDashStyle style, Color color, float width) => throw new Exception();
+        public void Initialize(PenDashStyle style, Color color, double width) => throw new Exception();
     }
 }
